ch3: refactor ch3 comm hints to use autogen init
