+++ src/mpid/ch3/src/mpid_init.c
@@ -34,7 +34,6 @@ char *MPIDI_DBG_parent_str = "?";
 static int init_pg(int *argc, char ***argv, int *has_parent, int *pg_rank_p, MPIDI_PG_t **pg_p);
 static int pg_compare_ids(void * id1, void * id2);
 static int pg_destroy(MPIDI_PG_t * pg );
-static int set_eager_threshold(MPIR_Comm *comm_ptr, MPIR_Info *info, void *state);
 
 MPIDI_Process_t MPIDI_Process = { NULL };
 MPIDI_CH3U_SRBuf_element_t * MPIDI_CH3U_SRBuf_pool = NULL;
@@ -64,28 +63,6 @@ static int finalize_failed_procs_group(void *param)
  fn_fail:
     return mpi_errno;
 }
-
-static int set_eager_threshold(MPIR_Comm *comm_ptr, MPIR_Info *info, void *state)
-{
-    int mpi_errno = MPI_SUCCESS;
-    char *endptr;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPIDI_CH3_SET_EAGER_THRESHOLD);
-
-    MPIR_FUNC_TERSE_ENTER(MPID_STATE_MPIDI_CH3_SET_EAGER_THRESHOLD);
-
-    comm_ptr->dev.eager_max_msg_sz = strtol(info->value, &endptr, 0);
-
-    MPIR_ERR_CHKANDJUMP1(*endptr, mpi_errno, MPI_ERR_ARG,
-                         "**infohintparse", "**infohintparse %s",
-                         info->key);
-
- fn_exit:
-    MPIR_FUNC_TERSE_EXIT(MPID_STATE_MPIDI_CH3_SET_EAGER_THRESHOLD);
-    return mpi_errno;
- fn_fail:
-    goto fn_exit;
-}
-
 
 int MPID_Init(int *argc, char ***argv, int requested, int *provided)
 {
@@ -267,11 +244,6 @@ int MPID_Init(int *argc, char ***argv, int requested, int *provided)
 	*provided = (MPICH_THREAD_LEVEL < requested) ? 
 	    MPICH_THREAD_LEVEL : requested;
     }
-
-    mpi_errno = MPIR_Comm_register_hint("eager_rendezvous_threshold",
-                                        set_eager_threshold,
-                                        NULL);
-    MPIR_ERR_CHECK(mpi_errno);
 
     mpi_errno = MPIDI_RMA_init();
     MPIR_ERR_CHECK(mpi_errno);
